Adds documentation to Rendering
Also a bit of cleanup
