+++ Orbis/Orbis.csproj
@@ -118,7 +118,7 @@
     </Compile>
     <Compile Include="Orbis.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Rendering\DecorationData.cs" />
+    <Compile Include="Rendering\DecorationPool.cs" />
     <Compile Include="UI\Elements\CivPanel.cs" />
     <Compile Include="UI\Elements\RelativeElement.cs" />
     <Compile Include="UI\IUpdateableElement.cs" />
